+++ CHANGELOG.md
@@ -29,6 +29,7 @@ Usage:
 * Fix #2397: Default JKube work directory (`jkube.workDir`) changed from `${project.build.directory}/jkube` to `${project.build.directory}/jkube-temp`
 * Fix #2393: Remove timestamp from `org.label-schema.build-date` LABEL to utilize docker cache
 * Fix #2399: Helm no longer generates default function; broadens support for different value types
+* Fix #2400: Helm supports complex values in `values.yaml` fragments (such as annotations or arrays)
 
 _**Note**_:
 - Container Images generated using jkube opinionated defaults no longer contain full timestamp in `org.label-schema.build-date` label. The label contains the build date in the format `yyyy-MM-dd`.